requirements.txt
google-api-python-client==2.154.0
google-auth==2.24.0
requests==2.31.0
# openai==1.40.8